+++ backend/src/main/java/com/carffeine/carffeine/station/domain/charger/Charger.java
@@ -1,6 +1,6 @@
 package com.carffeine.carffeine.station.domain.charger;
 
-import com.carffeine.carffeine.station.domain.BaseEntity;
+import com.carffeine.carffeine.common.domain.BaseEntity;
 import com.carffeine.carffeine.station.domain.station.Station;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
