+++ backend/src/main/java/com/carffeine/carffeine/station/domain/congestion/PeriodicCongestion.java
@@ -1,6 +1,6 @@
 package com.carffeine.carffeine.station.domain.congestion;
 
-import com.carffeine.carffeine.station.domain.BaseEntity;
+import com.carffeine.carffeine.common.domain.BaseEntity;
 import com.carffeine.carffeine.station.domain.charger.Charger;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
